+++ src/redux/modules/data.js
@@ -1,7 +1,26 @@
 import { normalize, denormalize } from 'normalizr';
 import currencies from '../../currencies';
-import query from '../../data/query';
 import apiSpec from '../../data/api-spec';
+import { mockGET } from '../../data/mock-api';
+import type { APISpec } from './api-spec';
+
+const getJSON = mockGET;
+
+const resolveURI = ({ uri }: APISpec, apiParams: ?Object): string =>
+  typeof uri === 'function' ? uri(apiParams || {}) : uri;
+
+const query = (
+  spec: APISpec,
+  apiParams: ?Object,
+  body: ?Object
+): Promise<*> => {
+  const uri = resolveURI(spec, apiParams);
+  if (spec.method === 'GET') {
+    return getJSON(uri);
+  }
+  throw new Error('no mock supported yet for method=' + spec.method);
+};
+
 
 // This initialize the initial entities (things like currencies are already available)
 const currenciesMap = {};
